?---4.12 is done.

+++ part4/bloglist/controllers/blogs.js
@@ -2,35 +2,69 @@ const express = require('express')
 const blogsRouter = express.Router()
 
 const Blog = require('../models/blogs')
+const User = require('../models/user')
+const config = require('../utils/config')
+const jwt = require('jsonwebtoken')
+
 
 
 blogsRouter.get('/', async (request, response) => {
     let allblogs = await Blog.find({})
+                    .populate('author', {userName:1, name:1})
     
     response.json(allblogs)
 })
 
 blogsRouter.post('/', async (request, response) => {
-    const blog = new Blog(request.body)
-    if(blog.likes === undefined){
-        blog.likes = 0
+    const body = request.body
+    const token = request.token
+    const decodedToken = jwt.verify(token, config.SECRET)
+
+    if(!token || !decodedToken.id){
+        return response.status(404).json({error: 'token missing or invalid.'})
     }
 
-    if(blog.url === undefined || blog.title === undefined){
+    
+    if(body.likes === undefined){
+        body.likes = 0
+    }
+
+    if(body.url === undefined || body.title === undefined){
         return response.status(400).end()
     }
 
+    const user = await User.findById(decodedToken.id)
+
+    body.author = user._id
+
+    const blog = new Blog(body)
     let savedBlog = await blog.save()
     response.status(201).json(savedBlog)
 })
 
+
 blogsRouter.delete('/:id', async (request, response) => {
     const id = request.params.id
+
+    const token = request.token
+    console.log(token)
+    const decodedToken = jwt.verify(token, config.SECRET)
+
+    if(!token || !decodedToken.id){
+        return response.status(403).json({error: 'no permition to access.'})
+    }
+
+    let blogToDel = await Blog.findById(id)
+    if(decodedToken.id.toString() !== blogToDel.author.toString()){
+        return response.status(403).json({error: 'no Permition to delete.'})
+    }
     
     let deleteRes = await Blog.deleteOne({_id: id})
     
     response.status(204).json(deleteRes)
 })
+
+
 blogsRouter.put('/:id', async (request, response) => {
     const id = request.params.id
     
@@ -38,5 +72,7 @@ blogsRouter.put('/:id', async (request, response) => {
     let updatedNote = await Blog.findById(id)
     response.json(updatedNote)
 })
+
+
 
 module.exports = blogsRouter
